refactor(store): updated store project to have same atributes than bigo project
