+++ projects/store/src/public/cards/gameCard/card.tsx
@@ -1,10 +1,7 @@
 import Text from "@4miga/design-system/components/Text";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import ImageNotFound from "public/img/ImageNotFound.jpg";
-import { useEffect, useState } from "react";
 import { ProductType } from "types/productTypes";
-import { checkImageUrl } from "utils/checkImageUrl";
 import { formatString } from "utils/formatString";
 import { CardContainer } from "./style";
 
@@ -14,16 +11,16 @@ interface CardProps {
 
 const GameCard = ({ product }: CardProps) => {
   const route = useRouter();
-  const [isImageValid, setIsImageValid] = useState<boolean>(false);
+  // const [isImageValid, setIsImageValid] = useState<boolean>(false);
 
-  useEffect(() => {
-    const checkImage = async () => {
-      const valid = await checkImageUrl(product.imgCardUrl);
-      setIsImageValid(valid);
-    };
+  // useEffect(() => {
+  //   const checkImage = async () => {
+  //     const valid = await checkImageUrl(product.imgCardUrl);
+  //     setIsImageValid(valid);
+  //   };
 
-    checkImage();
-  }, [product.imgCardUrl]);
+  //   checkImage();
+  // }, [product.imgCardUrl]);
 
   const handleProductClick = (path: string) => {
     const res = formatString(path);
@@ -33,7 +30,7 @@ const GameCard = ({ product }: CardProps) => {
   return (
     <CardContainer onClick={() => handleProductClick(product.name)}>
       <Image
-        src={isImageValid ? product.imgCardUrl : ImageNotFound}
+        src={product.imgCardUrl}
         alt={`Imagem do jogo ${product.name}`}
         height={200}
         width={200}
