+++ projects/store/src/public/components/carousel/carousel.tsx
