+++ projects/store/src/public/components/carousel/style.ts
