+++ projects/store/src/public/components/lines/style.ts
